Run user application in PID namespace
Launcher will put all it's children in PID namespace.
From now on the applications spawn by launer will not be
able to send signals to system daemons.

Currently launcher is creating only PID namespace. So
the getpid() and pid in /proc directory will not match.
This is the reason why a small patch to aul will be required.

Change-Id: If2998ef958b37bbb45e4e6c9af0a6fcd4f86496c